% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils_for_network_estimation.R
\name{conf_matrix}
\alias{conf_matrix}
\title{Confusion matrix for network estimation}
\usage{
conf_matrix(
  true_adj,
  est_adj,
  margins = FALSE,
  normalize = FALSE,
  undirected = TRUE
)
}
\arguments{
\item{true_adj}{The true \code{p} by \code{p} adjacency matrix of the network.}

\item{est_adj}{The estimated \code{p} by \code{p} adjacency matrix of the network.}

\item{margins}{Boolean, add marginal sums if \code{TRUE}. The default value is \code{FALSE}.}

\item{normalize}{Boolean, use normalized values (proportions) instead of absolute values if
\code{TRUE}. The default value is \code{FALSE}.}

\item{undirected}{Boolean, if \code{TRUE} (default) assumes that the adjacency matrices are
symmetric.}
}
\value{
A data frame, which size depends if marginal sums are displayed or not. The default size
 is 2 by 2, and if marginal sums added, then 3 by 3.
}
\description{
Construct confusion matrix using the true adjacency and estimated adjacency matrices of
the network. Can be used both directed and undirected networks.
}
\examples{
true_adj <- matrix(c(0, 0, 1, 0, 0,
                     0, 0, 1, 1, 0,
                     1, 1, 0, 0, 0,
                     0, 1, 0, 0, 1,
                     0, 0, 0, 1, 0), 5, 5)
est_adj <- matrix(c(0, 1, 1, 0, 0,
                    1, 0, 0, 1, 0,
                    1, 0, 0, 0, 1,
                    0, 1, 0, 0, 1,
                    0, 0, 1, 1, 0), 5, 5)
conf_matrix(true_adj, est_adj)
}
