% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_timeseries_network_data.R
\name{generate_timeseries_network_data}
\alias{generate_timeseries_network_data}
\title{Time series network data generator}
\usage{
generate_timeseries_network_data(
  n = 100,
  p = 100,
  n_timepoints = 10,
  n_edges = 0,
  n_datasets = 1,
  change_points = c(5),
  n_add_del = 5,
  rho = 0.25,
  w1 = 0.1,
  w2 = 0.3,
  smooth_precision = TRUE,
  seed_net = 12345,
  seed_data = 123456
)
}
\arguments{
\item{n}{The number of samples per time point. A scalar or a \code{n_timepoints}-long vector.}

\item{p}{The number of variables.}

\item{n_timepoints}{The number of time points.}

\item{n_edges}{The number of edges in the network. The default value is 0 meaning that number
of edges is set to \code{p}.}

\item{n_datasets}{The number of generated datasets, if multiple replicates needed. The default
value is 1.}

\item{change_points}{A vector containing time points where changes happen.}

\item{n_add_del}{The number of changing edges in each change point.}

\item{rho}{Initial value for the diagonal of the precision matrix.}

\item{w1}{Lower bound for the weights used to determine value of the precision matrix.}

\item{w2}{Upper bound for the weights used to determine value of the precision matrix.}

\item{smooth_precision}{If \code{TRUE}, generated precision matrices change smoothly over time.
If \code{FALSE}, changes happen immediately in the change points. The default value is \code{TRUE}.}

\item{seed_net}{Seed for network structure generation.}

\item{seed_data}{Seed for datasets generation.}
}
\value{
A list, which contains following objects:
\item{Sigmas}{
   A \code{p} by \code{p} by \code{n_timepoints} array containing simulated covariance matrices.
}
\item{Omegas}{
   A \code{p} by \code{p} by \code{n_timepoints} array containing simulated precision matrices.
}
\item{Thetas}{
   A \code{p} by \code{p} by \code{n_timepoints} array containing simulated adjacency matrices.
}
\item{datasets}{
   An \code{n_datasets}-long list, each containing \code{n_timepoints}-long list, where each
   list item is \code{n} by \code{p} matrix of the simulated data per time point.
}
}
\description{
This function generates time series network data with a random network structure
}
\examples{
sim <- generate_timeseries_network_data(n = 50, p = 100, n_timepoints = 20,
                                        change_points = c(5, 10, 15), n_add_del = 10)
}
