\name{rTVGL-package}
\alias{rTVGL-package}
\alias{rTVGL}
\docType{package}
\title{
\packageTitle{rTVGL}
}
\description{
\packageDescription{rTVGL}
}
\details{
%%  ~~ An overview of how to use the package, ~~
%%  ~~ including the most important functions ~~
}
\author{
\packageAuthor{rTVGL}

Maintainer: \packageMaintainer{rTVGL}
}
\references{
%%  ~~ Literature or other references for background information ~~
}
\keyword{package}

%% Uncomment below to imitate parts of library(help = rTVGL)
%\section{The \file{DESCRIPTION} File}{\packageDESCRIPTION{rTVGL}}
%\section{Documentation Index}{\packageIndices{rTVGL}}

\seealso{
%%  ~~ Optional links to other man pages, e.g. ~~
%%  ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%%  ~~ Optional simple examples of the most important functions ~~
}
