% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tvgl.R
\name{tvgl}
\alias{tvgl}
\title{Time-Varying Graphical LASSO for R}
\usage{
tvgl(
  data,
  lambda = 0,
  beta = 0,
  penalty_type = "l1",
  use_correlation = TRUE,
  rho = 1,
  tol = 1e-04,
  rtol = 1e-04,
  max_iterations = 1000,
  use_Cpp = TRUE,
  verbose = 0
)
}
\arguments{
\item{data}{A list or an array containing the data, where each list item or array "slice" is
\code{n} by \code{p} matrix, where \code{n} = sample size and \code{p} = number of variables.
If an array provided, time must be the last dimension of the array.}

\item{lambda}{Numeric, a regularization parameter that controls the sparsity of the networks.}

\item{beta}{Numeric, a regularization parameter that controls the how strongly adjacent time
points are pulled together.}

\item{penalty_type}{What kind of penalization is used over the time. Possible choices are
\code{"l1"} and \code{"laplacian"}. The default value is \code{"l1"}.}

\item{use_correlation}{If \code{TRUE}, the sample correlation is used instead of sample covariance
and number of observations is set to be 1. The default value is \code{TRUE}.}

\item{rho}{An Augmented Lagrangian penalty parameter in the ADMM algorithm.}

\item{tol}{An absolute tolerance for the convergence. The default value is 1e-4.}

\item{rtol}{A relative tolerance for the convergence. The default value is 1e-4.}

\item{max_iterations}{A maximum numbers of the ADMM algorithm iterations. The default value is 500.}

\item{use_Cpp}{Boolean, which is used to select the C++ or R version of the algorithm. The default
value is \code{TRUE} (C++ version).}

\item{verbose}{Numeric with possible values of 0, 1 and 2. 0 (default) prints only information of
final iteration. 1 prints information every 10th iteration, and 2 prints information every
iteration.}
}
\value{
A list, which contains following objects:
\item{Omega_ests}{
   A \code{p} by \code{p} by \code{n_timepoints} array containing the precision matrix estimates.
}
\item{Theta_ests}{
   A \code{p} by \code{p} by \code{n_timepoints} array containing adjacency matrices of the
   network estimates.
}
\item{iters}{
   The number of ADMM algorithm iterations.
}
\item{total_time}{
   The total run time of the algorithm in seconds.
}
}
\description{
This function performs the ADMM algorithm for time-varying graphical LASSO (TVGL or TVGLASSO)
introduced by Hallac et al. (2017) and adapted from the Python package \code{regain} 
(Tomasi et al., 2018).
}
\examples{
sim <- generate_timeseries_network_data(n = 50, p = 100, n_timepoints = 20,
                                        change_points = c(5, 10, 15), n_add_del = 10)
results <- tvgl(sim$datasets[[1]], lambda = 0.25, beta = 0.25, penalty_type = "l1")
}
\references{
Hallac, D., Park, Y., Boyd, S., & Leskovec, J. (2017). Network inference via the 
   time-varying graphical lasso. In \emph{Proceedings of the 23rd ACM SIGKDD international
   conference on knowledge discovery and data mining} (pp. 205-213).

Tomasi, F., Tozzo, V., Salzo, S., & Verri, A. (2018). Latent variable time-varying network
   inference. In \emph{Proceedings of the 24th ACM SIGKDD International Conference on Knowledge
   Discovery & Data Mining} (pp. 2338-2346).
}
